+++ lib/smurff-cpp/MatrixConfig.h
@@ -12,8 +12,8 @@ namespace smurff
    class MatrixConfig : public TensorConfig
    {
    private:
-      mutable std::shared_ptr<std::vector<std::uint64_t> > m_rows;
-      mutable std::shared_ptr<std::vector<std::uint64_t> > m_cols;
+      mutable std::shared_ptr<std::vector<std::uint32_t> > m_rows;
+      mutable std::shared_ptr<std::vector<std::uint32_t> > m_cols;
 
    //
    // Dense double matrix constructos
@@ -43,24 +43,24 @@ namespace smurff
    public:
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , const std::vector<std::uint64_t>& rows
-                  , const std::vector<std::uint64_t>& cols
+                  , const std::vector<std::uint32_t>& rows
+                  , const std::vector<std::uint32_t>& cols
                   , const std::vector<double>& values
                   , const NoiseConfig& noiseConfig
                   );
 
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , std::vector<std::uint64_t>&& rows
-                  , std::vector<std::uint64_t>&& cols
+                  , std::vector<std::uint32_t>&& rows
+                  , std::vector<std::uint32_t>&& cols
                   , std::vector<double>&& values
                   , const NoiseConfig& noiseConfig
                   );
 
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , std::shared_ptr<std::vector<std::uint64_t> > rows
-                  , std::shared_ptr<std::vector<std::uint64_t> > cols
+                  , std::shared_ptr<std::vector<std::uint32_t> > rows
+                  , std::shared_ptr<std::vector<std::uint32_t> > cols
                   , std::shared_ptr<std::vector<double> > values
                   , const NoiseConfig& noiseConfig
                   );
@@ -71,22 +71,22 @@ namespace smurff
    public:
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , const std::vector<std::uint64_t>& rows
-                  , const std::vector<std::uint64_t>& cols
+                  , const std::vector<std::uint32_t>& rows
+                  , const std::vector<std::uint32_t>& cols
                   , const NoiseConfig& noiseConfig
                   );
 
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , std::vector<std::uint64_t>&& rows
-                  , std::vector<std::uint64_t>&& cols
+                  , std::vector<std::uint32_t>&& rows
+                  , std::vector<std::uint32_t>&& cols
                   , const NoiseConfig& noiseConfig
                   );
 
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , std::shared_ptr<std::vector<std::uint64_t> > rows
-                  , std::shared_ptr<std::vector<std::uint64_t> > cols
+                  , std::shared_ptr<std::vector<std::uint32_t> > rows
+                  , std::shared_ptr<std::vector<std::uint32_t> > cols
                   , const NoiseConfig& noiseConfig
                   );
 
@@ -96,21 +96,21 @@ namespace smurff
    public:
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , const std::vector<std::uint64_t>& columns
+                  , const std::vector<std::uint32_t>& columns
                   , const std::vector<double>& values
                   , const NoiseConfig& noiseConfig
                   );
 
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , std::vector<std::uint64_t>&& columns
+                  , std::vector<std::uint32_t>&& columns
                   , std::vector<double>&& values
                   , const NoiseConfig& noiseConfig
                   );
 
       MatrixConfig( std::uint64_t nrow
                   , std::uint64_t ncol
-                  , std::shared_ptr<std::vector<std::uint64_t> > columns
+                  , std::shared_ptr<std::vector<std::uint32_t> > columns
                   , std::shared_ptr<std::vector<double> > values
                   , const NoiseConfig& noiseConfig
                   );
@@ -122,10 +122,10 @@ namespace smurff
       std::uint64_t getNRow() const;
       std::uint64_t getNCol() const;
 
-      const std::vector<std::uint64_t>& getRows() const;
-      const std::vector<std::uint64_t>& getCols() const;
+      const std::vector<std::uint32_t>& getRows() const;
+      const std::vector<std::uint32_t>& getCols() const;
 
-      std::shared_ptr<std::vector<std::uint64_t> > getRowsPtr() const;
-      std::shared_ptr<std::vector<std::uint64_t> > getColsPtr() const;
+      std::shared_ptr<std::vector<std::uint32_t> > getRowsPtr() const;
+      std::shared_ptr<std::vector<std::uint32_t> > getColsPtr() const;
    };
 }
